Add files via upload
Repeating ability description fixed.
Data is fetched and stored in a let variable (set to null), than the list of pokemon is created using the data within the variable. 
New & altered functions: 
fetchAbilityData - if statement included to check if the variable fetchedAbilityData is true and uses that data in the displayAbilityPokemon() function. 
fetchPokemonList
displayAbilityPokemon - .slice included to remove the ability description based on offest and limit