Update index.js
Notes:
To stop the repeating of the ability description a let variable was initially defined as null and later used to store the data fetched when searching for an ability. This stored data is then used to create the list of pokemon to display beneath the ability description. Slice is used to remove the ability description from showing on each scroll (cuts off the first integer).
Using another let variable, the program determines when all the available data has been displayed and stops it from repeating the results. It is a stop function to prevent unlimited results. The allPokemonFetched variable is set to true when the slicedPokemonList (pokemon list mapped from the abilityData variable minus the ability description) has a length equal to 0, and therefore no more data will be fetched as dictated within the scroll event listener.